Major files reorganisation
I changed most of the file location to improve the stability and
make it eaiser to work on code edior.

My hypothesis of the previous bug is that the program failed to
locate certain directory due to the additional languages selection
pages. So I fixed it by reduce the number of file hierrachy for
hopefully to make the website more streamlined.